% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_amazon_hls_2020}
\alias{samples_amazon_hls_2020}
\title{Land Cover in the Amazon Rainforest using HLSL30 for 2020}
\format{
A \code{sits} tibble containing 1,489 (463 already available) samples
extracted from HLSL30 (30 meters resolution) in Amazon Rainforest.
Each time-series covers one year of 3-month composites
(01/01/2020 to 12/31/2020).

Also, the \code{sits} tibble contains the following columns:
longitude (East-west coordinate of the time series sample in WGS 84),
latitude (North-south coordinate of the time series sample in WGS 84),
start_date (initial date of the time series),
end_date (final date of the time series),
label (the class label associated to the sample),
cube (the name of the cube from where the data was extracted), and
time_series (list containing a tibble with the values of the time series).

The training samples were collected by ground surveys
and high-resolution image interpretation by experts from
the Brazilian National Institute for Space Research (INPE)
team and partners. It has one class (\code{Forest}).

Each time series covers 12 months (4 data points) from the Landsat product,
and has six spectral bands (\code{BLUE}, \code{GREEN}, \code{RED},
\code{NIR08}, \code{SWIR16}, \code{SWIR22}) and four spectral indices
(\code{NDVI}, \code{EVI}, \code{MNDWI}, \code{NBR}).
}
\usage{
data("samples_amazon_hls_2020")
}
\description{
A dataset containing a sits tibble with time-series samples from Amazon
Rainforest in Brazil for 2020.
}
\keyword{datasets}
