% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_amazonbiome_landsat_2013}
\alias{samples_amazonbiome_landsat_2013}
\title{Land Use and Land Cover in Amazon Biome using LANDSAT-C2-L2 for 2013-2014}
\format{
A \code{sits} tibble containing 598 samples
extracted from LANDSAT-C2-L2 (30 meters resolution) in Amazon Biome
(Brazil). Each time-series covers one year of 3-month composites
(01/09/2013 to 01/06/2014).

Also, the \code{sits} tibble contains the following columns:
longitude (East-west coordinate of the time series sample in WGS 84),
latitude (North-south coordinate of the time series sample in WGS 84),
start_date (initial date of the time series),
end_date (final date of the time series),
label (the class label associated to the sample),
cube (the name of the cube from where the data was extracted), and
time_series (list containing a tibble with the values of the time series).

The training samples were collected by ground surveys
and high-resolution image interpretation by experts from
the Brazilian National Institute for Space Research (INPE)
team and partners. It has 5 classes land cover (\code{Pasture}, \code{Savanna},
\code{Soy_Fallow}, \code{Soy_Corn}, \code{Soy_Cotton}.

Each time series covers 12 months (4 data points) from the Landsat product,
and has six spectral bands (\code{BLUE}, \code{GREEN}, \code{RED},
\code{NIR08}, \code{SWIR16}, \code{SWIR22}) and four spectral indices
(\code{NDVI}, \code{EVI}, \code{MNDWI}, \code{NBR}).
}
\usage{
data("samples_amazonbiome_landsat_2013")
}
\description{
A dataset containing a sits tibble with time-series samples from Amazon Biome
in Brazil for 2013
}
\keyword{datasets}
