% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_baixotocantins_landsat_1996}
\alias{samples_baixotocantins_landsat_1996}
\title{LULC in Baixo Tocantins using LANDSAT-C2-L2 for 1996}
\format{
A \code{sits} tibble containing 465 samples extracted from
LANDSAT-C2-L2 (30 meters resolution) in Baixo Tocantins. Each time-series
covers one year of 3-month composites (01/01/1996 to 12/31/1996).

Also, the \code{sits} tibble contains the following columns:
longitude (East-west coordinate of the time series sample in WGS 84),
latitude (North-south coordinate of the time series sample in WGS 84),
start_date (initial date of the time series),
end_date (final date of the time series),
label (the class label associated to the sample),
cube (the name of the cube from where the data was extracted), and
time_series (list containing a tibble with the values of the time series).

The training samples were collected by ground surveys
and high-resolution image interpretation by experts from
the Brazilian National Institute for Space Research (INPE)
team and partners. It has 9 classes nine land cover
classes (\code{Small-Scale Agriculture - AGPE},
\code{All Water Bodies - AGUA}, \code{Primary Forest - FLO},
\code{Others - OT (Beaches, Outcrops, and Sandbanks)},
\code{Clean Pasture - PL},
\code{Dirty Pasture and Pasture with Regeneration - PSR},
\code{Urbanized Area - URB}, \code{Advanced Secondary Vegetation - VSA},
\code{Initial Secondary Vegetation - VSI}).

Each time series covers 12 months (4 data points) from the Landsat product,
and has six spectral bands (\code{BLUE}, \code{GREEN}, \code{RED},
\code{NIR08}, \code{SWIR16}, \code{SWIR22}) and four spectral indices
(\code{NDVI}, \code{EVI}, \code{MNDWI}, \code{NBR}).
}
\usage{
data("samples_baixotocantins_landsat_1996")
}
\description{
A dataset containing a sits tibble with time-series samples from Baixo
Tocantins in Brazil for 1996.
}
\keyword{datasets}
